moved osm converters to map_action_logic
